adding authorization to posts(protecting posts)

+++ controllers/posts.js
@@ -21,8 +21,8 @@ class Posts {
     post.save();
     return post;
   }
-  async removePost(id) {
-    return this.collection.deleteOne({ _id: ObjectId(id) });
+  async removePost(id, user_id) {
+    return this.collection.deleteOne({ _id: ObjectId(id), user_id });
   }
   async editPost(postUpdateFileds) {
     return this.collection.findOneAndUpdate(
@@ -31,8 +31,14 @@ class Posts {
       { new: true }
     );
   }
-  async removePostComments(id) {
-    return this.helperCollection.deleteMany({ post_id: ObjectId(id) });
+  async removePostComments(id, user_id) {
+    // user_id is the id of the person who comments
+    // const postOwner = 1;
+    console.log(id, user_id);
+    return this.helperCollection.deleteMany({
+      post_id: ObjectId(id),
+      user_id,
+    });
   }
 }
 module.exports = new Posts();
